pyproject.toml
[project]
name = "claude-x"
version = "0.6.0"
description = "Second Brain and Command Center for Claude Code"
readme = "README.md"
license = { text = "MIT" }
authors = [
    { name = "lucas.ms", email = "karismalms@naver.com" }
]
requires-python = ">=3.10"
dependencies = [
    "rich>=14.2.0",
    "typer>=0.21.1",
    "click>=8.1",
    "watchdog>=6.0.0",
    "pydantic>=2.12.5",
    "mcp>=1.0.0",
    "pyyaml>=6.0",
]
keywords = [
    "claude",
    "claude-code",
    "ai",
    "assistant",
    "prompt-engineering",
    "code-analysis",
    "session-tracking",
    "prompt-library",
    "mcp",
    "model-context-protocol",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Utilities",
]

[project.urls]
Homepage = "https://github.com/kakao/claude-x"
Documentation = "https://github.com/kakao/claude-x#readme"
Repository = "https://github.com/kakao/claude-x"
Issues = "https://github.com/kakao/claude-x/issues"
Changelog = "https://github.com/kakao/claude-x/blob/main/CHANGELOG.md"

[project.scripts]
cx = "claude_x.cli:main"
cx-mcp = "claude_x.mcp_server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/claude_x"]

[tool.hatch.build.targets.sdist]
include = [
    "/src",
]

[tool.hatch.build]
include = [
    "src/claude_x/**/*.yaml",
    "src/claude_x/**/*.yml",
]
